+++ interfacing/src/App/MQTT_APP/mqtt_app.cpp
@@ -1,7 +1,7 @@
 #include "mqtt_app.h"
 #include <Arduino.h>
 #include "../../Hal/MQTT/mqtt_core.h"
-#include "../../Hal/SoilMoisture/SoilMoisture.h"
+#include "../../App/SoilMoisture/SoilMoisture.h"
 #include "../../App/DHT/DHT11.h"
 #include "../../Hal/Pump/Pump.h"
 #include "../../Hal/WIFI/wifi.h"
@@ -75,10 +75,10 @@ void MQTT_APP_PublishTelemetry(void)
     SoilMoisture_getMoisture(&soilMoistureRaw);
     float soilMoisture = (float)soilMoistureRaw;
 
-    int temperature = 0;
+    float temperature = 0.0f;
     DHT11_GetTemperature(&temperature);
 
-    int humidity = 0;
+    float humidity = 0.0f;
     DHT11_GetHumidity(&humidity);
 
     // Create telemetry payload
@@ -86,7 +86,7 @@ void MQTT_APP_PublishTelemetry(void)
     telemetryPayload += "\"site\":\"site1\",";
     telemetryPayload += "\"node\":\"nodeA\",";
     telemetryPayload += "\"soil_moisture\":" + String(soilMoisture, 1) + ",";
-    telemetryPayload += "\"temperature\":" + String(temperature) + ",";
+    telemetryPayload += "\"temperature\":" + String((int)temperature) + ",";
     telemetryPayload += "\"humidity\":" + String(humidity);
     // Unimplemented sensors
     // telemetryPayload += ",\"ph\":0.0,";
